+++ src/abstract.cpp
@@ -175,7 +175,7 @@ namespace Genetic{
 			m_itercount++;
 		}
 	}
-	const unsigned long int AbstractGenetic::iteration_count()const{
+	const unsigned long long int AbstractGenetic::iteration_count()const{
 		return m_itercount;
 	}
 	const size_t AbstractGenetic::PopulationSize()const{
